requirements.txt
argon2-cffi==21.1.0
asgiref==3.5.2
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
bleach==4.1.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.3
dominate==2.6.0
entrypoints==0.3
filelock==3.8.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
gensim==4.2.0
imageio==2.19.3
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8.4
nltk==3.6.7
notebook==6.4.5
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pdfkit==1.0.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.4
prometheus-client==0.12.0
prompt-toolkit==3.0.22
pycparser==2.21
pygame==2.1.2
Pygments==2.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
pywin32==302
pywinpty==1.1.5
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
regex==2022.1.18
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==6.3.0
sqlparse==0.4.3
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2022.5.4
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tzdata==2022.7
virtualenv==20.16.7
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.6.0
